#21 Progress in the work on the building LAH

+++ src/main/java/ru/putnik/lahbuilder/link/AperiodicLink1.java
@@ -33,7 +33,7 @@ public class AperiodicLink1 extends Link{
     }
 
     @Override
-    public double getIncline() {
+    public int getIncline() {
         return -20;
     }
 }
