+++ src/main/java/ru/putnik/lahbuilder/link/DifferentialLink.java
@@ -29,7 +29,7 @@ public class DifferentialLink extends Link {
     }
 
     @Override
-    public double getIncline() {
+    public int getIncline() {
         return 0;
     }
 }
